+++ shareaza/DCNeighbour.h
@@ -36,7 +36,7 @@ public:
 
 	virtual BOOL	ConnectTo(const IN_ADDR* pAddress, WORD nPort, BOOL bAutomatic);
 	virtual BOOL	Send(CPacket* pPacket, BOOL bRelease = TRUE, BOOL bBuffered = FALSE);
-	virtual DWORD	GetUserCount() const { return m_oUsers.GetCount(); }
+	virtual DWORD	GetUserCount() const { return (DWORD)m_oUsers.GetCount(); }
 
 	// Process packets from input buffer
 	virtual BOOL	ProcessPackets(CBuffer* pInput);
